% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allPeptidesPlot.R
\name{allPeptidesPlot}
\alias{allPeptidesPlot}
\title{Create scatter plot
This function creates a scatter plot of all peptides that belong to a single proteins}
\usage{
allPeptidesPlot(
  env,
  dataFrame,
  max_frac,
  meanLine = FALSE,
  repPepLine = FALSE,
  separateLabStates = FALSE,
  grid = TRUE,
  titleLabel = "all",
  titleAlign = "left",
  ylabel = "Precursor Area",
  xlabel = "Fraction",
  legendLabel = "Condition",
  labelled = "Labeled",
  unlabelled = "Unlabeled"
)
}
\arguments{
\item{env:}{environtmnt, environtment containing data frames of peptides for each protein
indexed by `Protein Group Accessions`}

\item{dataFrame:}{character, `Protein Group Accession` to show in the plot}

\item{max_frac:}{numeric, total number of fractions}

\item{meanLine:}{logical, specifies whether to plot a mean line}

\item{repPepLine:}{logical, specifies whether to plot a representative peptide line}

\item{separateLabStates:}{logical, specifies whether label states will be separated into facets}

\item{titleAlign:}{character, one of the 'left', 'center'/'centre', 'right', specifies alignment of the title in plot}

\item{ylabel:}{character}

\item{xlabel:}{character}

\item{legendLabel:}{character}

\item{grid:}{logical, specifies presence/absence of gridline in the plot}

\item{labelled:}{character, label to be used for isLabel == TRUE}

\item{unlabelled:}{character, label to be used for isLabel == FALSE}
}
\description{
Create scatter plot
This function creates a scatter plot of all peptides that belong to a single proteins
}
