% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignClusters.R
\name{assignClusters}
\alias{assignClusters}
\title{Make a distmatrix}
\usage{
assignClusters(DF, distObject, method = "complete", cutoff = 0.5)
}
\arguments{
\item{DF}{DF: data frame, contains columns: `Protein Group Accessions` character `Protein Descriptions` character
isLabel character ('TRUE'/'FALSE')
               columns 1 to n, numeric, n is the total number of fractions/slices, each of this columns
                                          contains `Precursor Area` values in a given fraction(columns) for a protein(rows)}

\item{distObject}{input data}

\item{method}{character One of 'average', 'single' or 'complete' (default), specifies the linkage method to be used inside R hclust() function}

\item{cutoff}{numeric, specifies the h value in R cutree() function, height at which to 'cut the tree', everything with distance below this value is assigned into same cluster everything with larger distance is in a different cluster extreme possible values are 0 to 2 (might not be reached for all data sets)}
}
\value{
dataframe
}
\description{
This function modifies the input DF data frame by performing clustering on distObject
and adding a 'cluster' column to the DF
}
