% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{Clean raw complexomic data}
\usage{
cleanData(.data, fCol = "Search ID")
}
\arguments{
\item{.data}{dataframe}

\item{fCol}{character The column containing the fractions, e.g. "Search ID" (default)}
}
\value{
dataframe
}
\description{
Perform initial, mandatory, cleaning of data
Check presence of reguired columns
Function to process raw input data into format required for subsequent analysis
.data is a data frame containing raw input data
This function checks (not neccessarily in this order):
1) are all required columns present
2) are these columns in correct format
3) renames Sequence ID column to Fraction and converts values in this column from letters to numbers
4) reorders Protein Group Accessions containing multiple proteins
5) removes rows in which PSM Ambiguity == 'Rejected'
6) removes rows in which # Protein Groups == 0
7) removes rows in which Precursor Area is NA
8) removes cols that are not used at all
}
