% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compranApp.R
\name{compranApp}
\alias{compranApp}
\title{Execute the complexomics Shiny app}
\usage{
compranApp()
}
\value{
Shiny app
}
\description{
Execute the complexomics Shiny app
}
\examples{
##to start the shiny app associated with ComPrAn package run
compranApp()
}
