% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBarPlotClusterSummary.R
\name{makeBarPlotClusterSummary}
\alias{makeBarPlotClusterSummary}
\title{Title}
\usage{
makeBarPlotClusterSummary(df, name = "sample 1")
}
\arguments{
\item{df}{data frame, contains columns:
`Protein Group Accessions` character
`Protein Descriptions` character
 isLabel character ('TRUE'/'FALSE')
 columns 1 to n, numeric, n is the total number of fractions/slices, each of this columns
             contains `Precursor Area` values in a given fraction(columns) for a protein(rows)
 cluster integer}

\item{name}{character, specifies the name of the sample}
}
\value{
plot
}
\description{
Title
}
