% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDist.R
\name{makeDist}
\alias{makeDist}
\title{Make disstance matrix}
\usage{
makeDist(df, centered = FALSE)
}
\arguments{
\item{df}{data frame, contains columns:
`Protein Group Accessions` character
`Protein Descriptions` character
isLabel character ('TRUE'/'FALSE')
columns 1 to n, numeric, n is the total number of 
fractions/slices, each of this columns 
contains `Precursor Area` values in a given fraction(columns) 
for a protein(rows)}

\item{centered}{centered: logical,if TRUE return dist matrix based on 
centered Pearson correlation (uses R cor() function, fast)
,if FALSE return dist matrix based on uncentered Pearson
correlation (uses custom uncenteredCor() function, slow)}
}
\value{
matrix
}
\description{
This function calculates distance matrix for a data frame, column by column
requires uncenteredCor function to work
}
