% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEnv.R
\name{makeEnv}
\alias{makeEnv}
\title{Create a Populated Environment}
\usage{
makeEnv(.data)
}
\arguments{
\item{.data}{peptides dataframe}
}
\value{
An environment
}
\description{
split a large dataframe into an environment of dataframes
}
