% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeTable.R
\name{normalizeTable}
\alias{normalizeTable}
\title{Convert extractRepPeps output to a Matrix}
\usage{
normalizeTable(.data, applyNormalization = TRUE)
}
\arguments{
\item{.data}{a dataframe}

\item{applyNormalization}{logical apply normalization or not}
}
\value{
a matrix
}
\description{
Convert the dataframe as output from extractRepPeps() to matrix-like table
return normalized or raw values of Precursor Area, by default return
normalized values
}
