% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normTableForExport.R
\name{normTableForExport}
\alias{normTableForExport}
\title{Convert Normalized Dataframe to Export format}
\usage{
normTableForExport(labTab, unlabTab, comboTab)
}
\arguments{
\item{labTab}{a dataframe}

\item{unlabTab}{a dataframe}

\item{comboTab}{a dataframe}
}
\value{
a dataframe
}
\description{
This is a convenient function for plotting
}
