% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normTableWideToLong.R
\name{normTableWideToLong}
\alias{normTableWideToLong}
\title{Convert Normalized Dataframe To Long format}
\usage{
normTableWideToLong(labTab, unlabTab, comboTab)
}
\arguments{
\item{labTab}{a dataframe}

\item{unlabTab}{a dataframe}

\item{comboTab}{a dataframe}
}
\value{
a dataframe
}
\description{
This is a convenient function for plotting
}
