% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlyInOneLabelState_ENV.R
\name{onlyInOneLabelState_ENV}
\alias{onlyInOneLabelState_ENV}
\title{Report Proteins Present In Only One Label State}
\usage{
onlyInOneLabelState_ENV(.data)
}
\arguments{
\item{.data}{An environment containing dataframes}
}
\value{
a list with 3 items, each item is a vector containing names belonging to one of 3 groups
}
\description{
This function returns NAMES of proteins present in only labelled/only unlabelld or both label states
}
