% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptideImport.R
\name{peptideImport}
\alias{peptideImport}
\title{Import raw peptide complexomics data}
\usage{
peptideImport(inputFile)
}
\arguments{
\item{inputFile}{character}
}
\value{
dataframe
}
\description{
Check presence of required columns
inputFile is a character vector containing the location of peptide file
This function checks:
\itemize{
\item are all required columns present
\item are these columns in correct format
}
}
\examples{
 
##Use example peptide data set, read in data
inputFile <- system.file("extdata", "data.txt", package = "ComPrAn")
peptides <- peptideImport(inputFile)
}
\author{
Petra Palenikova \email{pp451@cam.ac.uk}

Rick Scavetta \email{office@scavetta.academy}
}
