% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protInportForAnalysis.R
\name{protInportForAnalysis}
\alias{protInportForAnalysis}
\title{Modify import protein data}
\usage{
protInportForAnalysis(.data)
}
\arguments{
\item{.data}{- data frame, contains columns:}
}
\value{
data frame
}
\description{
This function converts imported protien table into format compatible with 
downstream analysis
Imported file needs to contain following columns:
\itemize{
\item "Protein Group Accessions" - character/factor
\item "Protein Descriptions" - character
\item "scenario" - character/factor
\item "label" - logical
\item  columns "1" to "n" - numeric
}
}
\examples{

##Use example normalised proteins file
inputFile <- system.file("extdata", "dataNormProts.txt", package = "ComPrAn")
#read file in and change structure of table to required format
forAnalysis <- protInportForAnalysis(data.table::fread(inputFile))

}
