% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncenteredCor.R
\name{uncenteredCor}
\alias{uncenteredCor}
\title{Perform uncentered correlation}
\usage{
uncenteredCor(xx, yy)
}
\arguments{
\item{xx}{numeric vector}

\item{yy}{numeric vector}
}
\value{
vector
}
\description{
Perform uncentered correlation
}
