% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEnv.R
\name{makeEnv}
\alias{makeEnv}
\title{Create a Populated Environment}
\usage{
makeEnv(.data)
}
\arguments{
\item{.data}{peptides dataframe}
}
\value{
An environment
}
\description{
split a large dataframe into an environment of dataframes
}
\examples{
##Use example peptide data set, read in and clean data
inputFile <- system.file("extdata", "data.txt", package = "ComPrAn")
peptides <- cleanData(data.table::fread(inputFile), fCol = "Search ID")
## separate chemical modifications and labelling into separate columns
peptides <- splitModLab(peptides) 
## remove unneccessary columns, simplify rows
peptides <- simplifyProteins(peptides) 
##make environment
peptide_index <- makeEnv(peptides)

}
