% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protInportForAnalysis.R
\name{protInportForAnalysis}
\alias{protInportForAnalysis}
\title{Modify import protein data
This function converts imported protien table into format compatible with downstream analysis}
\usage{
protInportForAnalysis(.data)
}
\arguments{
\item{.data}{- data frame, contains columns:}

\item{"Protein}{Descriptions" - character}

\item{"scenario"}{- character/factor}

\item{"label"}{- logical}

\item{columns}{"1" to "n" - numeric}
}
\description{
Modify import protein data
This function converts imported protien table into format compatible with downstream analysis
}
